requirements.txt
opencv-python
numpy
Pillow
pytesseract
requests